requirements.txt
Pillow>=10.0.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
tqdm>=4.66.0
open3d>=0.17.0
folium>=0.14.0
reverse_geocoder>=1.5.1
